+++ bulk_grades/api.py
@@ -29,7 +29,7 @@ log = logging.getLogger(__name__)
 UNKNOWN_LAST_SCORE_OVERRIDER = 'unknown'
 
 
-def _get_enrollments(course_id, track=None, cohort=None):
+def _get_enrollments(course_id, track=None, cohort=None, active_only=False):
     """
     Return iterator of enrollment dictionaries.
 
@@ -51,6 +51,8 @@ def _get_enrollments(course_id, track=None, cohort=None):
         enrollments = enrollments.filter(
             user__cohortmembership__course_id=course_id,
             user__cohortmembership__course_user_group__name=cohort)
+    if active_only:
+        enrollments = enrollments.filter(is_active=True)
     for enrollment in enrollments:
         enrollment_dict = {
             'user': enrollment.user,
@@ -281,6 +283,7 @@ class GradeCSVProcessor(DeferrableMixin, GradedSubsectionMixin, CSVProcessor):
         self.track = None
         self.cohort = None
         self.user_id = None
+        self.active_only = False
 
         # The CSVProcessor.__init__ method will set attributes on self
         # from items in kwargs, so this super().__init__() call can
@@ -380,7 +383,7 @@ class GradeCSVProcessor(DeferrableMixin, GradedSubsectionMixin, CSVProcessor):
         """
         Return iterator of rows to export.
         """
-        enrollments = list(_get_enrollments(self._course_key, track=self.track, cohort=self.cohort))
+        enrollments = list(_get_enrollments(self._course_key, track=self.track, cohort=self.cohort, active_only=self.active_only))
         enrolled_users = [enroll['user'] for enroll in enrollments]
 
         grades_api.prefetch_course_and_subsection_grades(self._course_key, enrolled_users)