+++ bulk_grades/views.py
@@ -133,6 +133,7 @@ class GradeImportExport(GradeOnlyExport):
                                       if assignment_grade_min else None),
                 course_grade_min=(float(course_grade_min) if course_grade_min else None),
                 course_grade_max=(float(course_grade_max) if course_grade_max else None),
+                active_only=True,
             )